Simplifiy by removing On Demand Sync from Full Controller and Toggle
