--- com.vrcfury.vrcfury/Editor/VF/Service/OnDemandSyncService.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 72f941cf22ab438c8f7fad963ab2ffc3
-timeCreated: 1706916514
